refactor: rename vscode-service to vscode-apis
