+++ src/api.ts
@@ -14,14 +14,14 @@ import * as uri from 'vs/base/common/uri'
 import * as log from 'vs/platform/log/common/log'
 import * as telemetryUtils from 'vs/platform/telemetry/common/telemetryUtils'
 import * as searchExtHostTypes from 'vs/workbench/services/search/common/searchExtTypes'
-import createL10nApi from './vscode-services/l10n'
-import createLanguagesApi from './vscode-services/languages'
-import createCommandsApi from './vscode-services/commands'
-import createWorkspaceApi from './vscode-services/workspace'
-import createWindowApi, { TextTabInput } from './vscode-services/window'
-import createEnvApi from './vscode-services/env'
-import createDebugApi from './vscode-services/debug'
-import createExtensionsApi from './vscode-services/extensions'
+import createL10nApi from './vscode-apis/l10n'
+import createLanguagesApi from './vscode-apis/languages'
+import createCommandsApi from './vscode-apis/commands'
+import createWorkspaceApi from './vscode-apis/workspace'
+import createWindowApi, { TextTabInput } from './vscode-apis/window'
+import createEnvApi from './vscode-apis/env'
+import createDebugApi from './vscode-apis/debug'
+import createExtensionsApi from './vscode-apis/extensions'
 import { getDefaultExtension } from './default-extension'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
