+++ src/service-override/markers.ts
@@ -1,4 +1,4 @@
-import '../vscode-services/missing-services'
+import '../missing-services'
 import { IEditorOverrideServices } from 'vs/editor/standalone/browser/standaloneServices'
 import 'vs/workbench/contrib/markers/browser/markers.contribution'
