+++ src/service-override/theme.ts
@@ -1,4 +1,4 @@
-import '../vscode-services/missing-services'
+import '../missing-services'
 import { WorkbenchThemeService } from 'vs/workbench/services/themes/browser/workbenchThemeService'
 import { IThemeExtensionPoint } from 'vs/workbench/services/themes/common/workbenchThemeService'
 import { IEditorOverrideServices } from 'vs/editor/standalone/browser/standaloneServices'
