+++ src/vscode-apis/debug.ts
@@ -1,7 +1,7 @@
 import type * as vscode from 'vscode'
 import { DebugConfigurationProviderTriggerKind } from 'vs/workbench/contrib/debug/common/debug'
 import { IExtensionDescription } from 'vs/platform/extensions/common/extensions'
-import { getExtHostServices } from './extHost'
+import { getExtHostServices } from '../ext-host/extHost'
 
 export default function create (getExtension: () => IExtensionDescription): typeof vscode.debug {
   return {
